Refactored in order to make algorithm testable
